refactor: Home board refactoring
